split main script
