+++ sequana/scripts/main/samplesheet.py
@@ -1,0 +1,46 @@
+#  This file is part of Sequana software
+#
+#  Copyright (c) 2016-2020 - Sequana Development Team
+#
+#  Distributed under the terms of the 3-clause BSD license.
+#  The full license is in the LICENSE file, distributed with this software.
+#
+#  website: https://github.com/sequana/sequana
+#  documentation: http://sequana.readthedocs.io
+#
+##############################################################################
+import click
+import colorlog
+
+from sequana.iem import IEM
+
+from .utils import CONTEXT_SETTINGS
+
+
+logger = colorlog.getLogger(__name__)
+
+
+@click.command(context_settings=CONTEXT_SETTINGS)
+@click.argument("name", type=click.STRING)
+@click.option("--check", is_flag=True)
+@click.option("--extract-adapters", is_flag=True)
+@click.option("--quick-fix", is_flag=True)
+@click.option("--output", default=None)
+def samplesheet(**kwargs):
+    """Utilities to manipulate sample sheet"""
+    name = kwargs["name"]
+    if kwargs["check"]:
+        iem = IEM(name)
+        iem.validate()
+        logger.info("SampleSheet looks correct")
+    elif kwargs["extract_adapters"]:
+        iem = IEM(name)
+        iem.to_fasta()
+    elif kwargs["quick_fix"]:
+        iem = IEM(name, tryme=True)
+        if kwargs["output"]:
+            filename = kwargs["output"]
+        else:
+            filename = name + ".fixed"
+        logger.info("Saving fixed version in {}".format(filename))
+        iem.quick_fix(output_filename=filename)
